seperate api & impl
