+++ src/main/java/net/fabricmc/loom/api/LoomGradleExtensionAPI.java
@@ -32,7 +32,7 @@ import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.file.ConfigurableFileCollection;
 
-import net.fabricmc.loom.MixinAnnotationProcessorExtension;
+import net.fabricmc.loom.extension.MixinApExtension;
 import net.fabricmc.loom.api.decompilers.LoomDecompiler;
 import net.fabricmc.loom.configuration.ide.RunConfigSettings;
 import net.fabricmc.loom.configuration.processors.JarProcessor;
@@ -82,7 +82,7 @@ public interface LoomGradleExtensionAPI {
 
 	NamedDomainObjectContainer<RunConfigSettings> getRunConfigs();
 
-	void mixin(Action<MixinAnnotationProcessorExtension> action);
+	void mixin(Action<MixinApExtension> action);
 
 	void setCustomManifest(String customManifest);
