+++ src/main/java/net/fabricmc/loom/configuration/CompileConfiguration.java
@@ -33,7 +33,7 @@ import org.gradle.api.tasks.compile.JavaCompile;
 import org.gradle.api.tasks.javadoc.Javadoc;
 import org.gradle.jvm.tasks.Jar;
 
-import net.fabricmc.loom.MixinAnnotationProcessorExtension;
+import net.fabricmc.loom.extension.MixinApExtension;
 import net.fabricmc.loom.LoomGradleExtension;
 import net.fabricmc.loom.build.mixin.JavaApInvoker;
 import net.fabricmc.loom.build.mixin.KaptApInvoker;
@@ -144,7 +144,7 @@ public final class CompileConfiguration {
 			System.setProperty("log4j.skipJansi", "true");
 
 			project.getLogger().info("Configuring compiler arguments for Java");
-			MixinAnnotationProcessorExtension mixinApExtension = LoomGradleExtension.get(project).getMixinApExtension();
+			MixinApExtension mixinApExtension = LoomGradleExtension.get(project).getMixinApExtension();
 			mixinApExtension.init();
 
 			new JavaApInvoker(project).configureMixin();
