+++ src/main/java/net/fabricmc/loom/extension/LoomGradleExtensionApiImpl.java
@@ -36,7 +36,6 @@ import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.plugins.BasePluginConvention;
 
-import net.fabricmc.loom.MixinAnnotationProcessorExtension;
 import net.fabricmc.loom.api.decompilers.LoomDecompiler;
 import net.fabricmc.loom.api.LoomGradleExtensionAPI;
 import net.fabricmc.loom.configuration.ide.RunConfigSettings;
@@ -161,7 +160,7 @@ public abstract class LoomGradleExtensionApiImpl implements LoomGradleExtensionA
 	}
 
 	@Override
-	public void mixin(Action<MixinAnnotationProcessorExtension> action) {
+	public void mixin(Action<MixinApExtension> action) {
 		action.execute(getMixinApExtension());
 	}
 
@@ -180,7 +179,7 @@ public abstract class LoomGradleExtensionApiImpl implements LoomGradleExtensionA
 
 	protected abstract LoomFiles getFiles();
 
-	protected abstract MixinAnnotationProcessorExtension getMixinApExtension();
+	protected abstract MixinApExtension getMixinApExtension();
 
 	// This is here to ensure that LoomGradleExtensionApiImpl compiles without any unimplemented methods
 	private final class EnsureCompile extends LoomGradleExtensionApiImpl {
@@ -200,7 +199,7 @@ public abstract class LoomGradleExtensionApiImpl implements LoomGradleExtensionA
 		}
 
 		@Override
-		protected MixinAnnotationProcessorExtension getMixinApExtension() {
+		protected MixinApExtension getMixinApExtension() {
 			throw new RuntimeException("Yeah... something is really wrong");
 		}
 	}
